feat: migrate to ESM and add JSDoc with type definitions
